pyproject.toml
[project]
name = "mental-health-diagnostic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "unstructured-client>=0.42.12",
    "weasyprint>=68.1",
]
